Rename Exception to Exceptions.
