+++ example/upload.php
@@ -1,7 +1,7 @@
 <?php
 
 use Vimeo\Vimeo;
-use Vimeo\Exception\VimeoUploadException;
+use Vimeo\Exceptions\VimeoUploadException;
 
 /**
  *   Copyright 2013 Vimeo
