+++ src/Vimeo/Exceptions/VimeoUploadException.php
@@ -1,4 +1,4 @@
-<?php namespace Vimeo\Exception;
+<?php namespace Vimeo\Exceptions;
 
 /**
  * VimeoUploadException class for failure to upload to the server.
